requirements.txt
# Core AI / Graph
langgraph>=0.2.34
langchain-openai>=0.1.7
langchain-core>=0.2.0

# Data & ML
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Environment & utilities
python-dotenv>=1.0.0
typing-extensions>=4.9.0

# Web app & Visualization
streamlit>=1.37.0
plotly>=5.18.0

# Web crawler
requests>=2.31.0
beautifulsoup4>=4.12.2

# Optional (for export / Markdown reports)
markdown>=3.5.2

# Phase 2: Advanced Features
flask>=2.3.0
flask-cors>=4.0.0
kaleido>=0.2.1

# Production dependencies
gunicorn>=21.2.0